push

+++ lib/models/subscription_model/datum.dart
@@ -14,7 +14,7 @@ class Datum {
   final String? startDate;
   @JsonKey(name: 'expiry_date')
   final String? expiryDate;
-  final int? amount;
+  final double? amount;
   @JsonKey(name: 'payment_method')
   final String? paymentMethod;
   final String? account;
@@ -54,7 +54,7 @@ class Datum {
     String? status,
     String? startDate,
     String? expiryDate,
-    int? amount,
+    double? amount,
     String? paymentMethod,
     String? account,
     DateTime? createdAt,
